page restructure
