--- src/components/molecules/index.js
@@ -1,4 +1,0 @@
-import BottomNavigation from './BottomNavigation';
-import PageHeader from './PageHeader';
-
-export { PageHeader, BottomNavigation };
