+++ src/components/organisms/index.js
@@ -1,0 +1,4 @@
+import BottomNavigation from './BottomNavigation';
+import PageHeader from './PageHeader';
+
+export { BottomNavigation, PageHeader };
